+++ components/brave_ads/core/internal/account/confirmations/queue/queue_item/confirmation_queue_item_test_util.h
@@ -24,7 +24,7 @@ void SaveConfirmationQueueItems(
     const ConfirmationQueueItemList& confirmation_queue_items);
 
 void BuildAndSaveConfirmationQueueItems(const ConfirmationInfo& confirmation,
-                                        int count);
+                                        size_t count);
 
 }  // namespace test
